pyproject.toml
[project]
name = "aviator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.5.0",
    "flower>=2.0.1",
    "ipykernel>=6.29.5",
    "logfire[httpx]>=3.12.0",
    "matplotlib>=3.10.1",
    "numpy==1.23.2",
    "pandas>=2.1.0",
    "polars>=1.26.0",
    "prophet>=1.1.6",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "pydantic-ai[logfire]>=0.0.52",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "selenium>=4.29.0",
    "setuptools>=78.1.0",
    "talipp>=2.5.0",
]
